+++ src/DuneApp.cpp
@@ -1028,7 +1028,7 @@ void DuneApp::OnFileClose(MainWindow *window)
     delete window;
 #endif
     if (m_windows.size() == 0) {
-        for (size_t j = 0; j < m_filesToDelete.size(); j++)
+        for (long j = 0; j < m_filesToDelete.size(); j++)
             swRemoveFile(*m_filesToDelete[j]);
         swUploadCleanupPasswd(m_upload);
         Exit();
@@ -1046,7 +1046,7 @@ void DuneApp::OnFileExit()
     for (i = m_windows.first(); i; i = i->next()) {
         delete i->item();
     }
-    for (size_t j = 0; j < m_filesToDelete.size(); j++)
+    for (long j = 0; j < m_filesToDelete.size(); j++)
         swRemoveFile(*m_filesToDelete[j]);
     swUploadCleanupPasswd(m_upload);
     Exit();
@@ -1108,7 +1108,7 @@ void
 DuneApp::AddToFilesToDelete(char* string)
 {
     // test if string is already in list
-    for (size_t i = 0; i < m_filesToDelete.size(); i++)
+    for (long i = 0; i < m_filesToDelete.size(); i++)
        if (strcmp(string,(char*) m_filesToDelete[i])==0)
           return;
     m_filesToDelete.append(new MyString(string));
