+++ src/MFColorRGBA.cpp
@@ -88,7 +88,7 @@ MFColorRGBA::setSFValue(int index, float r, float g, float b, float a)
 void 
 MFColorRGBA::clamp(const FieldValue *min, const FieldValue *max)
 {
-    for (size_t i = 0; i < m_value.size(); i++) {
+    for (long i = 0; i < m_value.size(); i++) {
         if (m_value[i] < 0.0f) m_value[i] = 0.0f;
         if (m_value[i] > 1.0f) m_value[i] = 1.0f;
     }
