+++ src/MFNode.cpp
@@ -43,7 +43,7 @@ MFNode::MFNode(const MFNode &other)
 {
     m_value = new NodeList();
     if (other.m_value)
-        for (size_t i = 0; i < other.m_value->size(); i++) {
+        for (long i = 0; i < other.m_value->size(); i++) {
             m_value->set(i, other.m_value->get(i)->copy());
             m_value->get(i)->ref();
             if (!m_value->get(i)->isPROTO())
@@ -81,7 +81,7 @@ MFNode::MFNode(NodeList *value)
     m_value = new NodeList();
     if (value == NULL)
         return;
-    for (size_t i = 0; i < value->size(); i++) {
+    for (long i = 0; i < value->size(); i++) {
         Node *node = value->get(i);
         if (node != NULL) {
             node->ref();
@@ -93,7 +93,7 @@ MFNode::MFNode(NodeList *value)
 
 MFNode::~MFNode()
 {
-    for (size_t i = 0; i < m_value->size(); i++) {
+    for (long i = 0; i < m_value->size(); i++) {
         Node *node = m_value->get(i);
         if (node != NULL)
             node->unref();
@@ -288,7 +288,7 @@ MFNode::equals(const FieldValue *value) const
     if (value->getType() == MFNODE) {
         NodeList *v = ((MFNode *) value)->getValues();
         if (v->size() != m_value->size()) return false;
-        for (size_t i = 0; i < m_value->size(); i++)
+        for (long i = 0; i < m_value->size(); i++)
             if (v->get(i) != m_value->get(i)) return false;
         return true;
     }
@@ -299,7 +299,7 @@ FieldValue *
 MFNode::addNode(Node *node, int index) const
 {
     NodeList *list = new NodeList();
-    for (size_t i = 0; i < m_value->size(); i++) {
+    for (long i = 0; i < m_value->size(); i++) {
         list->append(m_value->get(i));
     }
     if (index == -1)
@@ -315,10 +315,10 @@ FieldValue *
 MFNode::removeNode(Node *node, int index) const
 {
     NodeList *list = new NodeList();
-    for (size_t i = 0; i < m_value->size(); i++) {
+    for (long i = 0; i < m_value->size(); i++) {
         list->append(m_value->get(i));
     }
-    for (size_t i = 0; i < list->size(); i++) {
+    for (long i = 0; i < list->size(); i++) {
         if (list->get(i) == node) {
             if (index == -1)
                 list->remove(i);
@@ -443,7 +443,7 @@ MFNode::preDraw()
 {
     NodeList *childList = getValues();
 
-    for (size_t i = 0; i < childList->size(); i++)
+    for (long i = 0; i < childList->size(); i++)
         childList->get(i)->preDraw();
 }
 
@@ -454,17 +454,17 @@ MFNode::draw(int pass, int mfNodeField)
 
     glPushName(mfNodeField);  // field offset
 
-    for (size_t i = 0; i < childList->size(); i++)
+    for (long i = 0; i < childList->size(); i++)
         childList->get(i)->bind();
 
     glPushName(0);
-    for (size_t i = 0; i < childList->size(); i++) {
+    for (long i = 0; i < childList->size(); i++) {
         glLoadName(i);
         childList->get(i)->draw(pass);
     }
     glPopName();
 
-    for (size_t i = 0; i < childList->size(); i++)
+    for (long i = 0; i < childList->size(); i++)
         childList->get(i)->unbind();
 
     glPopName();
@@ -477,7 +477,7 @@ MFNode::flip(int index)
 {
     NodeList *childList = getValues();
 
-    for (size_t i = 0; i < childList->size(); i++) {
+    for (long i = 0; i < childList->size(); i++) {
         Node *node = childList->get(i);
         if (!node->getFlag(NODE_FLAG_FLIPPED)) {
             node->setFlag(NODE_FLAG_FLIPPED);
@@ -491,7 +491,7 @@ MFNode::swap(int fromTo)
 {
     NodeList *childList = getValues();
 
-    for (size_t i = 0; i < childList->size(); i++) {
+    for (long i = 0; i < childList->size(); i++) {
         Node *node = childList->get(i);
         if (!node->getFlag(NODE_FLAG_SWAPPED)) {
             node->setFlag(NODE_FLAG_SWAPPED);
@@ -506,7 +506,7 @@ MFNode::countPolygons(void)
     int ret = 0;
     NodeList *childList = getValues();
 
-    for (size_t i = 0; i < childList->size(); i++)
+    for (long i = 0; i < childList->size(); i++)
         ret += childList->get(i)->countPolygons();
 
     return ret;
@@ -518,7 +518,7 @@ MFNode::countPrimitives(void)
     int ret = 0;
     NodeList *childList = getValues();
 
-    for (size_t i = 0; i < childList->size(); i++)
+    for (long i = 0; i < childList->size(); i++)
         ret += childList->get(i)->countPrimitives();
 
     return ret;
@@ -530,7 +530,7 @@ MFNode::countPolygons1Sided(void)
     int ret = 0;
     NodeList *childList = getValues();
 
-    for (size_t i = 0; i < childList->size(); i++)
+    for (long i = 0; i < childList->size(); i++)
         ret += childList->get(i)->countPolygons1Sided();
 
     return ret;
@@ -542,7 +542,7 @@ MFNode::countPolygons2Sided(void)
     int ret = 0;
     NodeList *childList = getValues();
 
-    for (size_t i = 0; i < childList->size(); i++)
+    for (long i = 0; i < childList->size(); i++)
         ret += childList->get(i)->countPolygons2Sided();
 
     return ret;
@@ -553,7 +553,7 @@ MFNode::update()
 {
     NodeList *childList = getValues();
 
-    for (size_t i = 0; i < childList->size(); i++)
+    for (long i = 0; i < childList->size(); i++)
         childList->get(i)->update();
 }
 
@@ -566,7 +566,7 @@ MFNode::convert2X3d(void)
     NodeList *newChilds = new NodeList();
 
     bool newNodes = false;
-    for (size_t i = 0; i < childList->size(); i++) {
+    for (long i = 0; i < childList->size(); i++) {
         if (childList->get(i) != NULL) {
             Node *newNode = childList->get(i)->convert2X3d();
             if (newNode != NULL) {
@@ -578,7 +578,7 @@ MFNode::convert2X3d(void)
     }
     if (newNodes) {            
         m_value = new NodeList();
-        for (size_t i = 0; i < newChilds->size(); i++) {
+        for (long i = 0; i < newChilds->size(); i++) {
             m_value->set(i, newChilds->get(i)->copy());
             m_value->get(i)->ref();
             if (!m_value->get(i)->isPROTO())
@@ -595,7 +595,7 @@ MFNode::convert2Vrml(void)
     NodeList *newChilds = new NodeList();
 
     bool newNodes = false;
-    for (size_t i = 0; i < childList->size(); i++) {
+    for (long i = 0; i < childList->size(); i++) {
         if (childList->get(i) != NULL) {
             Node *newNode = childList->get(i)->convert2Vrml();
             if (newNode != NULL) {
@@ -607,7 +607,7 @@ MFNode::convert2Vrml(void)
     }
     if (newNodes) {            
         m_value = new NodeList();
-        for (size_t i = 0; i < newChilds->size(); i++) {
+        for (long i = 0; i < newChilds->size(); i++) {
             m_value->set(i, newChilds->get(i)->copy());
             m_value->get(i)->ref();
             if (!m_value->get(i)->isPROTO())
@@ -658,7 +658,7 @@ MFNode::getBboxSize(void)
     Vec3f minResult(FLT_MAX, FLT_MAX, FLT_MAX);
     Vec3f maxResult(FLT_MIN, FLT_MIN, FLT_MIN);
     NodeList *childList = getValues();
-    for (size_t i = 0; i < childList->size(); i++)
+    for (long i = 0; i < childList->size(); i++)
         if (childList->get(i) != NULL) {
             Vec3f min = childList->get(i)->getMinBoundingBox();
             Vec3f max = childList->get(i)->getMaxBoundingBox();
@@ -697,7 +697,7 @@ MFNode::getBboxCenter(void)
     Vec3f minResult(FLT_MAX, FLT_MAX, FLT_MAX);
     Vec3f maxResult(FLT_MIN, FLT_MIN, FLT_MIN);
     NodeList *childList = getValues();
-    for (size_t i = 0; i < childList->size(); i++)
+    for (long i = 0; i < childList->size(); i++)
         if (childList->get(i) != NULL) {
             Vec3f min = childList->get(i)->getMinBoundingBox();
             Vec3f max = childList->get(i)->getMaxBoundingBox();
